Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	deleted:    .devcontainer/devcontainer.json
	new file:   Data_analysis_mid_240831_241110/Streamlit
	deleted:    README.md
	deleted:    requirements.txt
	deleted:    streamlit_app.py

Changes not staged for commit:
	modified:   Data_analysis_mid_240831_241110/Streamlit (modified content)

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.

Date:      Wed Nov 13 15:41:16 2024 +0900

On branch main
Your branch is ahead of 'origin/main' by 1 commit.
  (use "git push" to publish your local commits)

Changes to be committed:
	deleted:    .devcontainer/devcontainer.json
	new file:   Data_analysis_mid_240831_241110/Streamlit
	deleted:    README.md
	deleted:    requirements.txt
	deleted:    streamlit_app.py